refactor: extract shared report helpers into report_helpers module
Move the byte-identical private helpers _build_info_lines, _evidence_links,
_detail, and _one_line out of report.py and builtin_reporters.py into a new
termproof/report_helpers.py, and import them in both. Behavior-preserving;
rendered output is unchanged.
